requirements.txt
# Azure Communication Services
azure-communication-callingserver==1.0.0
azure-communication-identity==1.4.0
azure-communication-phonenumbers==1.1.0

# Azure Monitoring and Logging
azure-monitor-opentelemetry==1.2.0
azure-identity==1.15.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0

# FastAPI and Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# OpenAI for TTS
openai==1.10.0

# Audio Processing
pydub==0.25.1

# Async and Concurrency
aiofiles==23.2.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Utilities
python-dotenv==1.0.1
python-multipart==0.0.6